Add currency to DiscountCurve_ and move instruments to ycinstrument
- Add ccy_ member to DiscountCurve_ for currency awareness
- Move Deposit_/Swap_ declarations and implementations from
  yccalibration to ycinstrument.hpp/cpp where they belong
- Rename CalibrationYieldCurve_ to CalibratedYieldCurve_
- Add ccy parameter to CalibrateYieldCurve and NewDiscountPWLF
- Add } // namespace Dal closing comments to ycimp and yccomponent
