+++ dal/auto/MG_DiscountPWLF_v1_Write.inc
@@ -1,10 +1,11 @@
 // This file is auto-generated by machinist. Please don't modify it manually.
 namespace DiscountPWLF_v1
 {
-    void XWrite(Archive::Store_& dst, const String_& name, const Vector_<Date_>& knotDates, const Vector_<double>& leftVals, const Vector_<double>& rightVals, const Handle_<DiscountCurve_>& base) {
+    void XWrite(Archive::Store_& dst, const String_& name, const String_& ccy, const Vector_<Date_>& knotDates, const Vector_<double>& leftVals, const Vector_<double>& rightVals, const Handle_<DiscountCurve_>& base) {
         using namespace Archive::Utils;
         dst.SetType("DiscountPWLF_v1");
         SetOptional(dst, "name", name);
+        SetOptional(dst, "ccy", ccy);
         Set(dst, "knotDates", knotDates);
         Set(dst, "leftVals", leftVals);
         Set(dst, "rightVals", rightVals);
