+++ dal/curve/discount.cpp
@@ -7,5 +7,5 @@
 #include <dal/curve/discount.hpp>
 
 namespace Dal {
-    DiscountCurve_::DiscountCurve_(const String_ &name): YCComponent_("DiscountCurve", name) {}
+    DiscountCurve_::DiscountCurve_(const String_ &name, const String_& ccy): YCComponent_("DiscountCurve", name), ccy_(ccy) {}
 } // namespace Dal
